add Analyzer which writes jets vector for HH analysis
first version, based on diboson analyzer, but changing the writing of
jet vector to write directly the std::vector<JetType> object.
There is no need to reset and initialize the vector object because this
is done per event. To be able to write custom objects to a ROOT file,
dictionaries have to be generated. Therefore, the classses.h and
classes_def.xml files have been added to the src folder together
with the BuildFile.xml (RConfig.h error otherwise).